+++ src/main/kotlin/com/github/strindberg/emacsj/universal/UniversalArgumentDelegate.kt
@@ -17,6 +17,7 @@ import com.github.strindberg.emacsj.search.RestorableActionHandler
 import com.github.strindberg.emacsj.search.RestorableTypedActionHandler
 import com.github.strindberg.emacsj.space.ACTION_DELETE_SPACE
 import com.github.strindberg.emacsj.ui.CommonUI
+import com.github.strindberg.emacsj.word.ACTION_TRANSPOSE_WORDS
 import com.github.strindberg.emacsj.zap.ACTION_ZAP_BACKWARD_TO
 import com.github.strindberg.emacsj.zap.ACTION_ZAP_BACKWARD_UP_TO
 import com.github.strindberg.emacsj.zap.ACTION_ZAP_FORWARD_TO
@@ -47,6 +48,7 @@ internal val singleActions = setOf(
     ACTION_ZAP_BACKWARD_UP_TO,
     ACTION_DELETE_SPACE,
     ACTION_TRANSPOSE_LINES,
+    ACTION_TRANSPOSE_WORDS,
     ACTION_COPY_ABOVE_COMMAND,
 )